requirements.txt
fastapi
uvicorn[standard]
google-generativeai
requests
pydantic-settings
python-dotenv